pyproject.toml
[project]
name = "sas-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{name = "SAS Institute Inc."}]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[project.scripts]
app = "sas_mcp_server.main:main"
app-stdio = "sas_mcp_server.stdio_server:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.14.4",
]
